Crate evm_domain_runtime

Modules

api

opaque

Opaque types.

Structs

BaseFeeThreshold

BlockGasLimit

EVM block gas limit is set to maximum to allow all the transaction stored on Consensus chain.

BlockHashCount

BoundDivision

ChannelFeeModel

ChannelInitReservePortion

ChannelReserveFee

DefaultBaseFeePerGas

DefaultElasticity

DomainChainByteFee

DustRemovalHandler

DustRemovalHandler used to collect all the SSC dust left when the account is reaped.

EVMCurrencyAdapter

ExistentialDeposit

ExtrinsicStorageFees

FinalDomainTransactionByteFee

FindAuthorTruncated

GasLimitPovSizeRatio

HoldIdentifierWrapper

Hold identifier for balances for this runtime.

IntoRuntimeCall

MaxLocks

MaxReserves

MmrProofVerifier

OnXDMRewards

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PostOnlyBlockHash

PrecompilesValue

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SelfChainId

SessionKeys

StorageKeys

SuicideQuickClearLimit

TransactionConverter

TransactionWeightFee

TransporterEndpointId

Version

WeightPerGas

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

EXISTENTIAL_DEPOSIT

Computed as ED = Account data size * Price per byte, where Price per byte = Min Number of validators * Storage duration (years) * Storage cost per year Account data size (80 bytes) Min Number of redundant validators (10) - For a stable and redundant blockchain we need at least a certain number of full nodes/collators. Storage duration (1 year) - It is theoretically unlimited, accounts will stay around while the chain is alive. Storage cost per year of (12 * 1e-9 * 0.1 ) - SSD storage on cloud hosting costs about 0.1 USD per Gb per month

GAS_PER_SECOND

Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU).

RUNTIME_API_VERSIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_GAS

Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Functions

block_weights

extract_signer

maximum_block_length

Maximum block length for all dispatches. This is set to 3.75 MiB since consensus chain supports on 3.75 MiB for normal

maximum_domain_block_weight

The maximum domain block weight with 3.25 MiB as proof size Consensus allows 3.75 MiB but Fraud proof can carry extra size along with proof size So we set the proof size to 3.25 MiB

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balance of an account.

Balances

BalancesConfig

BaseFee

BaseFeeConfig

Block

Block type as expected by this runtime.

BlockFees

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

EVM

EVMChainId

EVMChainIdConfig

EVMConfig

EVMNoncetracker

Ethereum

EthereumConfig

Executive

Executive: handles dispatch to the various modules.

ExecutivePallet

Hash

A hash of some data used by the chain.

Messenger

Nonce

Index of a transaction in the chain.

Precompiles

Precompiles we use for EVM

SelfDomainId

SelfDomainIdConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

Sudo

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Transporter

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility